backend_py/user_grid3d_ri/pyproject.toml
[tool.poetry]
package-mode = false
name = "user-grid3d-ri"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = "^0.20.0"
httpx = "^0.24.0"
psutil = "^5.9.5"
pydantic = "^2.3.0"
numpy = "^1.24.1"
xtgeo = "^3.0.0"
diskcache = "^5.6.3" 
aiofiles = "^23.2.1"

# From ResInsight's requirements.txt
protobuf = "^4.25.3"
grpcio = "^1.62.1"
grpcio-tools = "^1.62.1"

[tool.black]
line-length = 120

